requirements.txt
flask
numpy
pandas
gunicorn
scikit-learn
Pillow
requests